backend/requirements.txt
django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
google-generativeai>=0.5.0
PyPDF2>=3.0.0
python-docx>=1.0.0
Pillow>=10.0.0
